separate tmbd and user rating

+++ app/src/main/java/com/example/a10xandroid/ui/journal/JournalModels.kt
@@ -20,7 +20,7 @@ data class JournalModelForView(
     val posterUrl: String?,
     val year: String,
     val genre: String,
-    val rating: Float,
+    val rating: Int,
     val addedAt: Long,
     val addedAtFormatted: String
 )
